refactor: renames button, separates path logic
