+++ src/views/product-tutorials-view/index.tsx
@@ -7,7 +7,7 @@ import ProductCollectionsSidebar, {
 import { ProductTutorialsSitemap } from './components'
 import { ProductTutorialsViewProps } from './server'
 import ProductViewContent from './components/product-view-content'
-import { OptOutButton } from './components/opt-out-button'
+import { OptOutButton } from './components/learn-opt-out-button'
 import { getOverviewHeading } from './helpers/heading-helpers'
 import s from './product-tutorials-view.module.css'
